move Header view into appropriate folder
