+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import Header from './Header';
+import Header from './views/Header';
 
 function App() {
   return (
